Add more documentation and re-structure some code
