+++ src/Cshrix/Cryptography/OneTimeKey.cs
@@ -6,11 +6,13 @@
 //   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 // </copyright>
 
-namespace Cshrix.Data
+namespace Cshrix.Cryptography
 {
     using System.Collections.Generic;
     using System.Collections.ObjectModel;
     using System.Linq;
+
+    using Data;
 
     using JetBrains.Annotations;
