+++ src/Cshrix/Data/Devices/DeviceMetadata.cs
@@ -1,4 +1,4 @@
-// <copyright file="SignedThirdPartyData.cs">
+// <copyright file="DeviceMetadata.cs">
 //   Copyright (c) 2019 by Adam Hellberg.
 //
 //   This Source Code Form is subject to the terms of the Mozilla Public
@@ -6,21 +6,18 @@
 //   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 // </copyright>
 
-namespace Cshrix.Data
+namespace Cshrix.Data.Devices
 {
-    using JetBrains.Annotations;
-
     using Newtonsoft.Json;
 
-    public readonly struct SignedThirdPartyData
+    public readonly struct DeviceMetadata
     {
         [JsonConstructor]
-        public SignedThirdPartyData([CanBeNull] SignedSenderData signed)
+        public DeviceMetadata(string displayName)
             : this() =>
-            Signed = signed;
+            DisplayName = displayName;
 
-        [CanBeNull]
-        [JsonProperty("third_party_signed")]
-        public SignedSenderData Signed { get; }
+        [JsonProperty("display_name")]
+        public string DisplayName { get; }
     }
 }
