+++ src/Cshrix/Data/Membership.cs
@@ -13,21 +13,39 @@ namespace Cshrix.Data
     using Newtonsoft.Json;
     using Newtonsoft.Json.Converters;
 
+    /// <summary>
+    /// Membership status in a room.
+    /// </summary>
     [JsonConverter(typeof(StringEnumConverter))]
     public enum Membership
     {
+        /// <summary>
+        /// User is invited to the room.
+        /// </summary>
         [EnumMember(Value = "invite")]
         Invited,
 
+        /// <summary>
+        /// User is joined to the room.
+        /// </summary>
         [EnumMember(Value = "join")]
         Joined,
 
+        /// <summary>
+        /// Reserved for future use.
+        /// </summary>
         [EnumMember(Value = "knock")]
         Knock,
 
+        /// <summary>
+        /// User has left the room.
+        /// </summary>
         [EnumMember(Value = "leave")]
         Left,
 
+        /// <summary>
+        /// User has been banned from the room.
+        /// </summary>
         [EnumMember(Value = "ban")]
         Banned
     }
