+++ src/Cshrix/Data/Presence/Presence.cs
@@ -6,22 +6,34 @@
 //   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 // </copyright>
 
-namespace Cshrix.Data
+namespace Cshrix.Data.Presence
 {
     using System.Runtime.Serialization;
 
     using Newtonsoft.Json;
     using Newtonsoft.Json.Converters;
 
+    /// <summary>
+    /// User presence.
+    /// </summary>
     [JsonConverter(typeof(StringEnumConverter))]
     public enum Presence
     {
+        /// <summary>
+        /// Unavailable for unknown reasons.
+        /// </summary>
         [EnumMember(Value = "unavailable")]
         Unavailable,
 
+        /// <summary>
+        /// User is currently online.
+        /// </summary>
         [EnumMember(Value = "online")]
         Online,
 
+        /// <summary>
+        /// User is offline.
+        /// </summary>
         [EnumMember(Value = "offline")]
         Offline
     }
