+++ src/Cshrix/Data/Presence/PresenceListUpdate.cs
@@ -6,7 +6,7 @@
 //   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 // </copyright>
 
-namespace Cshrix.Data
+namespace Cshrix.Data.Presence
 {
     using System.Collections.Generic;
     using System.Linq;
