+++ src/Cshrix/Data/Reason.cs
@@ -12,8 +12,16 @@ namespace Cshrix.Data
 
     using Newtonsoft.Json;
 
+    /// <summary>
+    /// Contains a reason for a kick or ban.
+    /// </summary>
     public readonly struct Reason
     {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Reason" /> structure.
+        /// </summary>
+        /// <param name="userId">The ID of the user affected.</param>
+        /// <param name="description">An optional reason description.</param>
         [JsonConstructor]
         public Reason(UserId userId, string description = null)
             : this()
@@ -22,14 +30,20 @@ namespace Cshrix.Data
             UserId = userId;
         }
 
+        /// <summary>
+        /// Gets the ID of the user whom this reason (kick or ban) applies to.
+        /// </summary>
+        [JsonProperty("user_id")]
+        public UserId UserId { get; }
+
+        /// <summary>
+        /// Gets a description of the reason.
+        /// </summary>
         [JsonProperty(
             "reason",
             NullValueHandling = NullValueHandling.Ignore,
             DefaultValueHandling = DefaultValueHandling.Ignore)]
         [CanBeNull]
         public string Description { get; }
-
-        [JsonProperty("user_id")]
-        public UserId UserId { get; }
     }
 }
