+++ test/Cshrix.Tests/Data/OneTimeKeyTests.cs
@@ -10,6 +10,8 @@ namespace Cshrix.Tests.Data
 {
     using System.Collections.Generic;
 
+    using Cryptography;
+
     using Cshrix.Data;
 
     using Newtonsoft.Json;
